fix(): restructures code into separate files
